Revert "Modify Project Workflows (#8)"
This reverts commit a964aa709e6fa84037e0c8c5d5b34ae42ce7a927.

+++ projects/aqlprofile/gfxip/gfx11/gfx11_primitives.h
@@ -639,7 +639,7 @@ class gfx11_cntx_prim {
 
   // Indicates the size of buffer to use per Shader Engine instance.
   // The size is specified in terms of 4KB blocks
-  static uint32_t sqtt_buffer_size_value(uint32_t size_val, uint32_t base_hi) {
+  static uint32_t sqtt_buffer_size_value(uint64_t size_val, uint32_t base_hi) {
 #if SQTT_PRIM_ENABLED
     uint32_t sq_thread_trace_buf0_size =
         SET_REG_FIELD_BITS(SQ_THREAD_TRACE_BUF0_SIZE, SIZE, size_val >> TT_BUFF_ALIGN_SHIFT) |
